requirements.txt
openpyxl
pandas
Pillow
pywin32
